+++ virttest/env_process.py
@@ -463,35 +463,10 @@ def preprocess(test, params, env):
                                                 image_name_only)
 
     # Start tcpdump if it isn't already running
-    if "address_cache" not in env:
-        env["address_cache"] = {}
-    if "tcpdump" in env and not env["tcpdump"].is_alive():
-        env["tcpdump"].close()
-        del env["tcpdump"]
-    if "tcpdump" not in env and params.get("run_tcpdump", "yes") == "yes":
-        cmd = "%s -npvi any 'port 68'" % utils_misc.find_command("tcpdump")
-        if params.get("remote_preprocess") == "yes":
-            login_cmd = ("ssh -o UserKnownHostsFile=/dev/null -o \
-                         PreferredAuthentications=password -p %s %s@%s" %
-                         (port, username, address))
-            env["tcpdump"] = aexpect.ShellSession(
-                login_cmd,
-                output_func=_update_address_cache,
-                output_params=(env["address_cache"],))
-            remote.handle_prompts(env["tcpdump"], username, password, prompt)
-            env["tcpdump"].sendline(cmd)
-        else:
-            env["tcpdump"] = aexpect.Tail(
-                command=cmd,
-                output_func=_tcpdump_handler,
-                output_params=(env["address_cache"], "tcpdump.log",))
-
-        if utils_misc.wait_for(lambda: not env["tcpdump"].is_alive(),
-                               0.1, 0.1, 1.0):
-            logging.warn("Could not start tcpdump")
-            logging.warn("Status: %s" % env["tcpdump"].get_status())
-            logging.warn("Output:" + utils_misc.format_str_for_message(
-                env["tcpdump"].get_output()))
+    # This thread will be automatically handled by the environment object,
+    # no need to explicitly close it. The fact it has to be started here
+    # is so that the test params have to be honored.
+    env.start_tcpdump(params)
 
     # Destroy and remove VMs that are no longer needed in the environment
     requested_vms = params.objects("vms")
@@ -776,11 +751,6 @@ def postprocess(test, params, env):
                 except Exception:
                     pass
 
-    # Terminate tcpdump if no VMs are alive
-    if not living_vms and "tcpdump" in env:
-        env["tcpdump"].close()
-        del env["tcpdump"]
-
     if params.get("setup_hugepages") == "yes":
         try:
             h = test_setup.HugePageConfig(params)
@@ -874,60 +844,6 @@ def postprocess_on_error(test, params, env):
     params.update(params.object_params("on_error"))
 
 
-def _update_address_cache(address_cache, line):
-    if re.search("Your.IP", line, re.IGNORECASE):
-        matches = re.findall(r"\d*\.\d*\.\d*\.\d*", line)
-        if matches:
-            address_cache["last_seen"] = matches[0]
-
-    if re.search("Client.Ethernet.Address", line, re.IGNORECASE):
-        matches = re.findall(r"\w*:\w*:\w*:\w*:\w*:\w*", line)
-        if matches and address_cache.get("last_seen"):
-            mac_address = matches[0].lower()
-            last_time = address_cache.get("time_%s" % mac_address, 0)
-            last_ip = address_cache.get("last_seen")
-            cached_ip = address_cache.get(mac_address)
-
-            if (time.time() - last_time > 5 or cached_ip != last_ip):
-                logging.debug("(address cache) DHCP lease OK: %s --> %s",
-                              mac_address, address_cache.get("last_seen"))
-
-            address_cache[mac_address] = address_cache.get("last_seen")
-            address_cache["time_%s" % mac_address] = time.time()
-            del address_cache["last_seen"]
-        elif matches:
-            address_cache["last_seen_mac"] = matches[0]
-
-    if re.search("Requested.IP", line, re.IGNORECASE):
-        matches = matches = re.findall(r"\d*\.\d*\.\d*\.\d*", line)
-        if matches and address_cache.get("last_seen_mac"):
-            ip_address = matches[0]
-            mac_address = address_cache.get("last_seen_mac")
-            last_time = address_cache.get("time_%s" % mac_address, 0)
-
-            if time.time() - last_time > 10:
-                logging.debug("(address cache) DHCP lease OK: %s --> %s",
-                              mac_address, ip_address)
-
-            address_cache[mac_address] = ip_address
-            address_cache["time_%s" % mac_address] = time.time()
-            del address_cache["last_seen_mac"]
-
-
-def _tcpdump_handler(address_cache, filename, line):
-    """
-    Helper for handler tcpdump output.
-
-    :params address_cache: address cache path.
-    :params filename: Log file name for tcpdump message.
-    :params line: Tcpdump output message.
-    """
-    try:
-        utils_misc.log_line(filename, line)
-    except Exception, reason:
-        logging.warn("Can't log tcpdump output, '%s'", reason)
-
-    _update_address_cache(address_cache, line)
 
 
 def _take_screendumps(test, params, env):
